+++ infra/postgres/init/01-create-test-database.sql
@@ -1,0 +1,1 @@
+CREATE DATABASE personalfin_test;